Rework the auto-generated protobuf file webdoc.
Put the file into its own subdirectory within the docs-src/content
directory as the index file so that it appears as the only page
when clicking on the reference link in the navigation menu.

Also updated some formatting and wording for the page. Added the
C++ grpc types for reference as well.